Addressing first round of comments
